rewrite o'clock
what started as a cleanup of peer code ballooned out of control into this.

most of the peer-related code now uses messages, reducing the amount of locks everywhere and allowing tasks to work freely.

a bunch of jank was also removed or replaced with different jank.

ui took a cue from the anime industry this season and got delayed to a later date