engine: fix status update handling
`.unwrap_or_default()`ing resulted in peers not being dropped when the watch channel was already closed. this resulted in some pieces being perpetually assigned to peers that were no longer active.

as a bonus, added validation code for piece assignment that releases pieces if the peer they are assigned to was dropped.